requirements.txt
streamlit==1.36.0
python-dotenv==1.0.1
requests==2.32.3
pandas
scikit-learn
sentence-transformers